Makefile improvements, also more file management.
